Remove dependency upper bounds, bump version
Most upper bounds have no justification, but even if there is
it may not affect some use cases. For example, the current
blocker for pyarrow v16 support is due to an HDFS issue. Anyone
who does not use HDFS as an artifact back-end will not care about
this issue. So there's no need to prevent installing pyarrow v16 with the
latest MLFlow in such cases.